[GEOT-7641] Support the "Sweep" parameter in the geostationary projections (#4896)
* Support the Geostationary projection with the GOES (x) sweep axis.
A fix to the geostationary projection in geotools to allow setting the sweep
axis as appropriate for GOES. The WKT parser only allows parameters
that can be parsed as doubles so I didn't add the sweep_axis as a parameter
with values x,y (which doesn't seem to be done anywhere anyway). We could fix
that, but it's a larger change. Instead, I added a sweep parameter that's
either a 0 (x) or a 1 (y).

All the math is copied over from PROJ of course, as it was before, but now it's
copied from a newer version of PROJ (5.2 instead of 4.9).

* Reformat